Move main.py inside of src/
